backend/requirements.txt
absl-py==2.1.0
aenum==3.1.15
aiohappyeyeballs==2.4.3
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
astunparse==1.6.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
Deprecated==1.2.14
distro==1.9.0
Django==5.1.2
django-jazzmin==3.0.1
django-stubs==5.1.0
django-stubs-ext==5.1.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
jiter==0.6.1
joblib==1.4.2
kiwisolver==1.4.7
libclang==18.1.1
line-bot-sdk==3.13.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
multidict==6.1.0
namex==0.0.8
numpy==1.26.4
openai==1.52.0
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
propcache==0.2.0
protobuf==4.25.5
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
pythainlp==5.0.4
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.2
safetensors==0.4.5
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
termcolor==2.5.0
threadpoolctl==3.5.0
tqdm==4.66.5
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
yarl==1.15.4
